rewrite restaurant logic

+++ src/app/Http/Controllers/Admin/RestaurantController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\Restaurant\StoreRestaurantRequest;
 use App\Http\Requests\Admin\Restaurant\UpdateRestaurantRequest;
-use App\Http\Resources\Admin\Restaurant\MinifiedRestaurantResource;
 use App\Models\Restaurant as RestaurantModel;
 use App\Facades\RestaurantFacade as Restaurant;
 use Illuminate\Http\RedirectResponse;
@@ -16,8 +15,8 @@ class RestaurantController extends Controller
 {
     public function index(): View
     {
-        $restaurants = RestaurantModel::query()->get();
-        $restaurants = MinifiedRestaurantResource::collection($restaurants)->toArray(\request());
+        $this->authorize('viewAny', RestaurantModel::class);
+        $restaurants = Restaurant::get();
         return view('admin.restaurants.index', compact('restaurants'));
     }
 
@@ -38,22 +37,18 @@ class RestaurantController extends Controller
 
     public function store(StoreRestaurantRequest $request): RedirectResponse
     {
-        Restaurant::store($request);
-        return redirect()
-            ->route('admin.restaurants.index')
-            ->with('success', __('Restaurant created successfully'));
+        $this->authorize('createNew', RestaurantModel::class);
+        return Restaurant::store($request);
     }
 
     public function update(UpdateRestaurantRequest $request, RestaurantModel $restaurant): RedirectResponse
     {
-        Restaurant::update($request, $restaurant);
-        return redirect()
-            ->route('admin.restaurants.index')
-            ->with('success', __('Restaurant updated successfully'));
+        return Restaurant::update($request, $restaurant);
     }
 
     public function destroy(RestaurantModel $restaurant): RedirectResponse
     {
+        $this->authorize('delete', $restaurant);
         $restaurant->delete();
         return back()
             ->with('success', __('Restaurant removed successfully'));
